Django MongoDB/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
Django==3.2.25
django-environ==0.12.0
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.7.0
mongoengine==0.29.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pymongo==3.11.4
pytz==2025.2
sqlparse==0.2.4
tzdata==2025.2
